backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-Compress==1.20

# Database
psycopg2-binary==2.9.11
SQLAlchemy==2.0.36

# Validation & Serialization
Marshmallow==3.20.1
python-dotenv==1.0.0

# Blockchain & Web3
web3==6.11.2
eth-keys==0.5.1

# Caching
redis==4.6.0

# WebSockets (Real-time updates)
flask-socketio==5.3.6
python-socketio==5.11.0

# File Upload
Werkzeug==3.0.1
boto3==1.34.7

# Security
bcrypt==4.1.1
PyJWT==2.10.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
setuptools>=65.5.1
PyGithub==2.1.1
openai==1.54.5
httpx<0.28.0  # Compatible with openai 1.54.5 (0.28+ breaks with 'proxies' arg)

# Email (Optional)
Flask-Mail==0.9.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production
gunicorn==21.2.0
gevent==24.2.1
gevent-websocket==0.10.1

# Task Queue & Background Jobs
celery[redis]==5.3.4

# QR Code Generation (for SBT verification)
qrcode[pil]==7.4.2
Pillow==10.2.0
